+++ hammer-api/src/main/java/cn/featherfly/hammer/expression/query/QuerySingleExecutor.java
@@ -9,19 +9,28 @@ import cn.featherfly.common.repository.mapping.RowMapper;
  * <p>
  * dsl for query single executor
  * </p>
+ * .
  *
  * @author zhongj
  */
 public interface QuerySingleExecutor {
+
     /**
-     * query for single
+     * query for single, return null when not found.
      *
      * @return map
      */
     Map<String, Object> single();
 
     /**
-     * query for single
+     * query for unique, throw JdbcException when not found.
+     *
+     * @return map
+     */
+    Map<String, Object> unique();
+
+    /**
+     * query for single, return null when not found.
      *
      * @param <E>  wrapper type
      * @param type wrapper type
@@ -30,11 +39,29 @@ public interface QuerySingleExecutor {
     <E> E single(Class<E> type);
 
     /**
-     * query for single
+     * query for unique, throw JdbcException when not found.
+     *
+     * @param <E>  the element type
+     * @param type the type
+     * @return the e
+     */
+    <E> E unique(Class<E> type);
+
+    /**
+     * query for single, return null when not found.
      *
      * @param <E>       wrapper type
      * @param rowMapper rowMapper
      * @return object
      */
     <E> E single(RowMapper<E> rowMapper);
+
+    /**
+     * query for unique, throw JdbcException when not found.
+     *
+     * @param <E>       the element type
+     * @param rowMapper the row mapper
+     * @return the e
+     */
+    <E> E unique(RowMapper<E> rowMapper);
 }