+++ hammer-sqldb/src/main/java/cn/featherfly/hammer/sqldb/jdbc/operate/MergeOperate.java
@@ -3,7 +3,7 @@ package cn.featherfly.hammer.sqldb.jdbc.operate;
 import java.sql.PreparedStatement;
 import java.util.Map;
 
-import com.speedment.common.tuple.Tuple2;
+import com.speedment.common.tuple.Tuple3;
 
 import cn.featherfly.common.db.mapping.ClassMappingUtils;
 import cn.featherfly.common.db.mapping.SqlTypeMappingManager;
@@ -72,13 +72,17 @@ public class MergeOperate<T> extends AbstractOperate<T> {
      * @return 操作影响的数据行数
      */
     public int execute(final T entity, boolean onlyNull) {
-        Tuple2<String, Map<Integer, String>> tuple = ClassMappingUtils.getMergeSqlAndParamPositions(entity,
+        Tuple3<String, Map<Integer, String>, Integer> tuple = ClassMappingUtils.getMergeSqlAndParamPositions(entity,
                 classMapping, onlyNull, jdbc.getDialect());
+        // 如果需要更新的参数数量为0,表示不需要更新
+        if (tuple.get2() == 0) {
+            return 0;
+        }
         return jdbc.execute((con, manager) -> {
             try (PreparedStatement prep = con.prepareStatement(tuple.get0())) {
                 Object[] params = setParameters(entity, tuple.get1(), prep, manager);
                 if (logger.isDebugEnabled()) {
-                    logger.debug("execute sql: {} \n params: {}", sql, ArrayUtils.toString(params));
+                    logger.debug("execute sql: {} \n params: {}", tuple.get0(), ArrayUtils.toString(params));
                 }
                 int result = prep.executeUpdate();
                 return result;